requirements.txt
absl-py==0.11.0
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
audioread==2.1.9
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.34.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.1.1
joblib==0.17.0
kapre==0.3.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
librosa==0.8.0
llvmlite==0.35.0
Markdown==3.3.3
matplotlib==3.3.3
numba==0.52.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.7
pandas==1.1.4
Pillow==8.0.1
pkg-resources==0.0.0
pooch==1.3.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.6
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
SoundFile==0.10.3.post1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-addons==0.11.2
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
typeguard==2.10.0
urllib3==1.26.2
Werkzeug==1.0.1
wget==3.2
wrapt==1.12.1
zipp==3.4.0
